Simple sql hooks

Made headway on SQLite logic

Commit while I backport a fix to 7.0

Mini-demo now works

Changes from review with Joe

Don't overuse flags-bits

Comment

Make sequence a specially handled type on the server side

Update logic working

llmeta functions

clang

master columns functions utilize llmeta, deadlock handling

First stab at schemachange

Working

cleanup a bit

Start nextsequence test

Test verifies basic functionality

Add prevent-modify tests

Prevent adding next nextsequence column which doesnt allow nulls

Modify dbstore value on insert as well as update, complete fastinit test

clang

Fix delete table crash

Fix comdb2sys and auth tests

remove unneeded code

Change from Rivers

formatting changes

Coerce ENOMEM errors in bdb_lite_fetch_keys_int to DBNOTFOUND

Add a property to comdb2_columns


Disable small sequences by default, better error messages, other review changes

Fix comdb2sys test


fix ddl_no_csc2

clang

Simple changes to fix tagged api

clang

Incorporate feedback from Joe and Akshat

WIP

Return correct error message for select nextsequence

Review comments from Nirbhay, Akshat, Rivers, and others

Print explicit error messages (thanks Nirbhay), checkin to work on docs

Regenerated bubble diagrams

Fix comdb2sys and auth tests

Changes from Nirbhay

Update nextsequence in docs

fix documentation

clang

Fix ddl_no_csc2 test

+++ schemachange/sc_drop_table.c
@@ -93,6 +93,11 @@ int finalize_drop_table(struct ireq *iq, struct schema_change_type *s,
 
     s->already_finalized = 1;
 
+    if ((rc = delete_table_sequences(tran, db))) {
+        sc_errf(s, "Failed deleting table sequences rc %d\n", rc);
+        return rc;
+    }
+
     delete_table(db, tran);
     /*Now that we don't have any data, please clear unwanted schemas.*/
     bdberr = bdb_reset_csc2_version(tran, db->tablename, db->schema_version);
